requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
djangorestframework==3.12.4
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
uuid==1.30
django-debug-toolbar==3.2.2
django-ckeditor==6.1.0
django-filter==2.4.0